+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added support for Redis PHP extension.
 - Added `REDIS_PARAMETER_LIMIT` env variable to avoid errors when calling Redis functions with large arrays. remp/remp#1307
 - Added index to `created_at` and `updated_at` campaign columns. remp/remp#1286
+- Fixed sorting campaigns by `is active` column. remp/helpdesk#2231
 
 ### [Mailer]